feat: pkg/dblib
